feat: crud issue endpoint

+++ src/services/issue-service/index.js
@@ -1,8 +1,9 @@
 const IssueRepository = require('../../repositories/issue-repository');
+const UserRepository = require('../../repositories/user-repository');
 const { catchThrows } = require('../../utils/promise');
 
 const { ImageIssueStorage } = require('../../utils/storage');
-const { issueStatus } = require('./constant');
+const { issueStatus, issueStaffRoles } = require('./constant');
 
 class IssueService {
     static async __uploadImage(images) {
@@ -26,6 +27,20 @@ class IssueService {
         );
     }
 
+    static async __accessValidation(oldIssue, userId) {
+        const user = await UserRepository.getUserById(userId);
+
+        const userRole = user.role;
+        if (
+            issueStaffRoles.includes(userRole) ||
+            oldIssue.user_creator_id === userId
+        ) {
+            return true;
+        }
+
+        return false;
+    }
+
     static async createIssue(data, files, userId) {
         const images = files.image || [];
         const uploadedImages = await this.__uploadImage(images);
@@ -47,8 +62,15 @@ class IssueService {
         };
     }
 
-    static async getIssue(id) {
+    static async getIssue(id, userId) {
         const issue = await IssueRepository.getIssue(id);
+
+        if (!(await this.__accessValidation(issue, userId))) {
+            return {
+                error_message:
+                    'Anda tidak memiliki akses untuk melihat laporan',
+            };
+        }
 
         return {
             message: 'Laporan berhasil didapatkan',
@@ -56,14 +78,22 @@ class IssueService {
         };
     }
 
-    static async updateIssue(id, data, files) {
+    static async updateIssue(id, data, files, userId) {
         const oldIssue = await IssueRepository.getIssue(id);
+
+        if (!(await this.__accessValidation(oldIssue, userId))) {
+            return {
+                error_message:
+                    'Anda tidak memiliki akses untuk mengubah laporan',
+            };
+        }
 
         const images = files.image || [];
         const uploadedImages = await this.__uploadImage(images);
 
         const issueData = {
             id,
+            user_creator_id: oldIssue.user_creator_id,
             user_assigned_id:
                 data.user_assigned_id || oldIssue.user_assigned_id,
             title: data.title || oldIssue.title,
@@ -81,8 +111,15 @@ class IssueService {
         };
     }
 
-    static async deleteIssue(id) {
+    static async deleteIssue(id, userId) {
         const issue = await IssueRepository.getIssue(id);
+
+        if (issue.user_creator_id !== userId) {
+            return {
+                error_message:
+                    'Anda tidak memiliki akses untuk menghapus laporan',
+            };
+        }
 
         await IssueRepository.deleteIssue(id);
         catchThrows(this.__deleteImage(issue.image));
